# api/docs/swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "This is the API for my personal dev website",
		"title": "Dev Website API",
		"termsOfService": "http://swagger.io/terms/",
		"contact": {
			"name": "Matheo Galuba",
			"url": "https://dev.matheo-galuba.com/contact",
			"email": "matheo.galu56@gmail.com"
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0.html"
		},
		"version": "1.0"
	},
	"host": "localhost:8080",
	"basePath": "/api",
	"paths": {
...
		"/ormi": {
			"get": {
				"security": [
					{
						"Bearer": []
					}
				],
				"description": "List a page of the todos of the authenticated user, filtered by state, labels and date of last state change, along with the total number of matches",
				"produces": ["application/json"],
				"tags": ["ormi"],
				"summary": "List todos",
				"parameters": [
					{
						"type": "array",
						"items": {
							"enum": ["TODO", "IN_PROGRESS", "STANDBY", "DONE", "CANCELLED"],
							"type": "string"
						},
						"collectionFormat": "multi",
						"description": "Only return todos in these states (repeat the parameter for several)",
						"name": "state",
						"in": "query"
					},
					{
						"type": "array",
						"items": {
							"type": "string"
						},
						"collectionFormat": "multi",
						"description": "Only return todos having all these labels (repeat the parameter for several)",
						"name": "label",
						"in": "query"
					},
					{
						"type": "string",
						"format": "date-time",
						"description": "Only return todos whose last state change is at or after this date",
						"name": "from",
						"in": "query"
					},
					{
						"type": "string",
						"format": "date-time",
						"description": "Only return todos whose last state change is before this date",
						"name": "to",
						"in": "query"
					},
					{
						"enum": ["position", "due", "created", "updated"],
						"type": "string",
						"default": "position",
						"description": "Custom order, due date (todos without one last), creation date or last state change",
						"name": "sort",
						"in": "query"
					},
					{
						"enum": ["asc", "desc"],
						"type": "string",
						"description": "Sort direction, defaults to desc for created and updated, asc otherwise",
						"name": "order",
						"in": "query"
					},
					{
						"maximum": 200,
						"minimum": 1,
						"type": "integer",
						"default": 50,
						"description": "Maximum number of results",
						"name": "limit",
						"in": "query"
					},
					{
						"minimum": 0,
						"type": "integer",
						"default": 0,
						"description": "Number of results to skip",
						"name": "offset",
						"in": "query"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/models.TodoPage"
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					}
				}
			},
			"post": {
				"security": [
					{
						"Bearer": []
					}
				],
				"description": "Create a todo in the TODO state, placed at the end of the list",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"tags": ["ormi"],
				"summary": "Create a todo",
				"parameters": [
					{
						"description": "Todo data",
						"name": "todo",
						"in": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/controllers.TodoBody"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Created",
						"schema": {
							"$ref": "#/definitions/models.Todo"
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					}
				}
			}
		},
		"/ormi/labels": {
			"get": {
				"security": [
					{
						"Bearer": []
					}
				],
				"description": "List the labels already used by the authenticated user, most used first, optionally restricted to those containing a text (case-insensitive)",
				"produces": ["application/json"],
				"tags": ["ormi"],
				"summary": "List used labels",
				"parameters": [
					{
						"type": "string",
						"description": "Text the labels must contain",
						"name": "q",
						"in": "query"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					}
				}
			}
		},
		"/ormi/order": {
			"put": {
				"security": [
					{
						"Bearer": []
					}
				],
				"description": "Set the position of the given todos to their index in the list",
				"consumes": ["application/json"],
				"tags": ["ormi"],
				"summary": "Reorder todos",
				"parameters": [
					{
						"description": "Todo IDs in their new order",
						"name": "order",
						"in": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/controllers.TodoOrderBody"
						}
					}
				],
				"responses": {
					"204": {
						"description": "No Content"
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					}
				}
			}
		},
		"/ormi/stats": {
			"get": {
				"security": [
					{
						"Bearer": []
					}
				],
				"description": "Todos completed per day over the last year (days without completion are omitted), completion streaks and counts of open todos",
				"produces": ["application/json"],
				"tags": ["ormi"],
				"summary": "Get todo stats",
				"parameters": [
					{
						"type": "string",
						"default": "UTC",
						"description": "IANA timezone used to split days",
						"name": "tz",
						"in": "query"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/models.TodoStats"
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					}
				}
			}
		},
		"/ormi/stats/{year}": {
			"get": {
				"security": [
					{
						"Bearer": []
					}
				],
				"description": "Count the todos completed per day of a calendar year, days without completion are omitted",
				"produces": ["application/json"],
				"tags": ["ormi"],
				"summary": "Get completions of a year",
				"parameters": [
					{
						"type": "integer",
						"description": "Calendar year",
						"name": "year",
						"in": "path",
						"required": true
					},
					{
						"type": "string",
						"default": "UTC",
						"description": "IANA timezone used to split days",
						"name": "tz",
						"in": "query"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/models.TodoDayCount"
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					}
				}
			}
		},
		"/ormi/{id}": {
			"get": {
				"security": [
					{
						"Bearer": []
					}
				],
				"description": "Get a todo of the authenticated user by id",
				"produces": ["application/json"],
				"tags": ["ormi"],
				"summary": "Get one todo",
				"parameters": [
					{
						"type": "string",
						"description": "Todo ID",
						"name": "id",
						"in": "path",
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/models.Todo"
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"404": {
						"description": "Not Found",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					}
				}
			},
			"delete": {
				"security": [
					{
						"Bearer": []
					}
				],
				"description": "Delete a todo of the authenticated user by id",
				"tags": ["ormi"],
				"summary": "Delete a todo",
				"parameters": [
					{
						"type": "string",
						"description": "Todo ID",
						"name": "id",
						"in": "path",
						"required": true
					}
				],
				"responses": {
					"204": {
						"description": "No Content"
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"404": {
						"description": "Not Found",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					}
				}
			},
			"patch": {
				"security": [
					{
						"Bearer": []
					}
				],
				"description": "Update the given fields of a todo. A null dueDate removes the due date, a new state is appended to the history.",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"tags": ["ormi"],
				"summary": "Update a todo",
				"parameters": [
					{
						"type": "string",
						"description": "Todo ID",
						"name": "id",
						"in": "path",
						"required": true
					},
					{
						"description": "Fields to update",
						"name": "todo",
						"in": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/controllers.PatchTodoBody"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/models.Todo"
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"404": {
						"description": "Not Found",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					}
				}
			}
		},
...
	},
	"definitions": {
...
		"controllers.PatchTodoBody": {
			"type": "object",
			"properties": {
				"description": {
					"type": "string",
					"maxLength": 1000
				},
				"dueDate": {
					"type": "string",
					"format": "date-time"
				},
				"gitIssue": {
					"type": "integer",
					"minimum": 1
				},
				"gitURL": {
					"type": "string"
				},
				"labels": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"state": {
					"type": "string",
					"enum": ["TODO", "IN_PROGRESS", "STANDBY", "DONE", "CANCELLED"]
				},
				"title": {
					"type": "string",
					"maxLength": 100,
					"minLength": 2
				}
			}
		},
...
		"controllers.TodoBody": {
			"type": "object",
			"required": ["title"],
			"properties": {
				"description": {
					"type": "string",
					"maxLength": 1000
				},
				"dueDate": {
					"type": "string",
					"format": "date-time"
				},
				"gitIssue": {
					"type": "integer",
					"minimum": 1
				},
				"gitURL": {
					"type": "string"
				},
				"labels": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"title": {
					"type": "string",
					"maxLength": 100,
					"minLength": 2
				}
			}
		},
		"controllers.TodoOrderBody": {
			"type": "object",
			"required": ["ids"],
			"properties": {
				"ids": {
					"type": "array",
					"maxItems": 1000,
					"minItems": 1,
					"items": {
						"type": "string"
					}
				}
			}
		},
...
		"models.OrmiEvent": {
			"type": "object",
			"properties": {
				"newState": {
					"type": "string",
					"example": "DONE"
				},
				"previousState": {
					"type": "string",
					"example": "TODO"
				},
				"updatedAt": {
					"type": "string",
					"format": "date-time"
				}
			}
		},
...
		"models.Todo": {
			"type": "object",
			"properties": {
				"createdAt": {
					"type": "string",
					"format": "date-time"
				},
				"description": {
					"type": "string"
				},
				"dueDate": {
					"type": "string",
					"format": "date-time"
				},
				"gitIssue": {
					"type": "integer"
				},
				"gitURL": {
					"type": "string"
				},
				"history": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/models.OrmiEvent"
					}
				},
				"id": {
					"type": "string"
				},
				"labels": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"position": {
					"type": "integer"
				},
				"state": {
					"type": "string",
					"enum": ["TODO", "IN_PROGRESS", "STANDBY", "DONE", "CANCELLED"]
				},
				"title": {
					"type": "string"
				},
				"user": {
					"type": "string"
				}
			}
		},
		"models.TodoDayCount": {
			"type": "object",
			"properties": {
				"count": {
					"type": "integer"
				},
				"date": {
					"type": "string",
					"example": "2026-10-09"
				}
			}
		},
		"models.TodoPage": {
			"type": "object",
			"properties": {
				"items": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/models.Todo"
					}
				},
				"total": {
					"type": "integer",
					"example": 42
				}
			}
		},
		"models.TodoStats": {
			"type": "object",
			"properties": {
				"currentStreak": {
					"type": "integer"
				},
				"days": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/models.TodoDayCount"
					}
				},
				"firstYear": {
					"type": "integer",
					"example": 2024
				},
				"inProgress": {
					"type": "integer"
				},
				"longestStreak": {
					"type": "integer"
				},
				"overdue": {
					"type": "integer"
				},
				"standby": {
					"type": "integer"
				},
				"todo": {
					"type": "integer"
				}
			}
		},
...
	},
	"securityDefinitions": {
		"Bearer": {
			"description": "Bearer token",
			"type": "apiKey",
			"name": "Authorization",
			"in": "header"
		}
	}
}